src/requirements.txt
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
requests==2.32.3

